Fix + Audit + CleanUp
git-svn-id: https://cchlib.googlecode.com/svn/trunk@8214 bc589049-105d-1923-a3ce-6fcb24364ed1

+++ cchlib-core/src/test/java/com/googlecode/cchlib/lang/reflect/MyTestByName.java
@@ -14,7 +14,7 @@ public class MyTestByName
 
     public String myTest( Integer value1, int value2 )
     {
-        return Integer.toString( value1 + value2 );
+        return Integer.toString( value1.intValue() + value2 );
     }
 
     public String myTest( int...values )
